+++ src/txio.C
@@ -167,7 +167,7 @@ void io_ns::print_ascii(ostream &fp, uchar c) {
 
 // returns true if only ascii 7 bits in the buffer
 auto Buffer::is_all_ascii() const -> bool {
-    for (int i = 0; i < wptr; i++) {
+    for (size_t i = 0; i < wptr; i++) {
         uchar c = at(i);
         if (c >= 128) return false;
         if (c < 32 && c != '\t' && c != '\n') return false;
@@ -178,7 +178,7 @@ auto Buffer::is_all_ascii() const -> bool {
 // returns false if some unprintable characters appear
 // Non-ascii chars are printable (assumes buffer is valid UTF8).
 auto Buffer::is_good_ascii() const -> bool {
-    for (int i = 0; i < wptr; i++) {
+    for (size_t i = 0; i < wptr; i++) {
         uchar c = at(i);
         if (c < 32 && c != '\t' && c != '\n') return false;
     }
@@ -236,7 +236,7 @@ void Buffer::utf8_error(bool first) {
                          << (first ? ", first byte" : ", continuation byte") << ")\n";
     main_ns::log_and_tty.L << "Position in line is " << ptr << lg_end;
     if (T.new_error()) return; // signal only one error per line
-    for (int i = 0; i < wptr; i++) io_ns::print_ascii(*(the_log.fp), at(i));
+    for (size_t i = 0; i < wptr; i++) io_ns::print_ascii(*(the_log.fp), at(i));
     the_log << lg_end;
 }
